Move utils.graphql to graphql package
