+++ src/services/workflow.service.js
@@ -19,7 +19,7 @@ import { GQuery } from '@/services/gquery'
 import store from '@/store/'
 import Alert from '@/model/Alert.model'
 
-import { createApolloClient } from '@/utils/graphql'
+import { createApolloClient } from '@/graphql'
 import { HOLD_WORKFLOW, RELEASE_WORKFLOW, STOP_WORKFLOW } from '@/graphql/queries'
 /* eslint-disable no-unused-vars */
 import { SubscriptionClient } from 'subscriptions-transport-ws'
